fix minor PR issues

+++ poll/services/users.py
@@ -1,5 +1,6 @@
 from poll.db.model_users import UserRepository
 from poll.schemas.users import SignUpReq, UserUpdateRes
+from poll.services.exc.user import UserNotFound
 
 
 class UserCRUD:
@@ -10,13 +11,20 @@ class UserCRUD:
         return await self.user_repository.get_all_users(page, page_size)
 
     async def get_user_by_id(self, user_id: int):
+        if not await self.user_repository.get_user_by_id(user_id):
+            raise UserNotFound(user_id)
         return await self.user_repository.get_user_by_id(user_id)
 
     async def create_user(self, user: SignUpReq):
         return await self.user_repository.create_user(user)
 
     async def update_user(self, user_id: int, user_update: UserUpdateRes):
-        return await self.user_repository.update_user(user_id, user_update)
+        user = await self.user_repository.get_user_by_id(user_id)
+        if user is None:
+            raise UserNotFound(user_id)
+        return await self.user_repository.update_user(user, user_update)
 
     async def delete_user(self, user_id: int):
+        if not await self.user_repository.get_user_by_id(user_id):
+            raise UserNotFound(user_id)
         await self.user_repository.delete_user(user_id)
